Automatically remove all dependent resources when the CrateDB resource is deleted
The approach taken in this commit makes use of Kubernetes' garbage
collection. We're setting an `OwnerReference` on each resource, including
the data PVC for the CrateDB containers. Since the `OwnerReference`
blocks the deletion of the owner resource, Kubernetes will trigger a
foreground deletion by default. That way, PVCs are also deleted
automatically.

https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/