Improve logging while waiting for cluster to be healthy again
Previously, nothing was logged when either the expected number of nodes
failed or the health check failed. This led to hard to debug problems if
e.g. a node was missing because of a discrepancy between the expected
number of nodes in the custom resource and the number of replicas in the
StatefulSets.