Use UUID4s for namespace names during tests
Previously, during tests, the namespace name was taken from the
`Faker.domain_word()` generator. However, when a name was reused "too
quickly", after the previous namespace was terminated, a new cluster
might have been created in the terminated namespace before that one is
actually gone.

What that means is, when triggering a namespace deletion, the nested
resources need to be deleted first. During that time, creating new
resources inside the namespace is still possible.

Now, if the creation is triggered successfully be the operator, and then
Kubernetes removes the custom resource and other resources, the
`create_handler` in the operator would still wait for the bootstrapping
of the cluster to succeed, which will never happen, because the cluster
isn't around anymore.