Ref #48: Make update handler more resilient
The operator used to have a long-running update handler. The handler
takes care of updating, restarting, and scaling a cluster. However, due
to the way the handler was implemented, it wasn't particular resilient.
Operator restarts would require the whole handler to be executed again.
This would cause all individual update tasks to be performed again as
well. Especially with cluster restarts, that could cause a lot of
problems, as nodes were restarted again, after having been restarted
already.

By using Kopf's sub-handler concept, individual sub-handlers are only
executed once. However, in order to be able to order the sub-handler, a
bit of state-handling is necessary. This commit implements the
state-handling and allows for arbitrary sub-handlers to be run in an
arbitrary order.