Make tests more resilient
When the operator deploys a CrateDB cluster, the StatefulSet adds a
local PersistentVolume to each pod which can be used for debugging.
Historically, the operator used the "default" storage class for these
PVs. However, since "default" may map to block storages in the cloud,
e.g. on Azure or AWS, K8s would attempt to mount the same block device
onto multiple pods, which isn't possible. To mitigate the issue, the
operator's tests now ensure there's a local file system.