Close the K8s API client's connections
Previously, when instantiating any of the kubernetes-asyncio APIs, such
as `CoreV1Api`, we didn't pass in an explicit `api_client` instance,
causing a new one to be created each time, which then would never close
the connection.

To change that, the `ApiClient` is only ever instantiated inside
`main.yaml` and instances of the `AppsV1Api`, `CoreV1Api`, etc. are
passed to functions. The only exception is the `WsApiClient` which is
instantiated when needed to keep the load on the K8s cluster low, by
closing those long-running connections right away after use.