Ref #48: Make create handler more resilient
Historically, the create handler for clusters was implemented in a
best-effort and with a bit of a "it'll be fine" approach. Kubernetes
resources were created and when they already existed that was fine. When
the create handler failed, though, the whole handler was being retried,
including the creation of already existing resources.

Now, with this change, only the individual sub-handlers will be retried,
making the process more efficient. The handler is now reentrant.

Unlike attempted in 07720e3080047c765211f6c56d70000b8b3a5043, the
ApiClient is now instantiated inside the individual sub-handlers, to
ensure a connection is available and open. Otherwise, sub-handlers will
fail with a `RuntimeError`: "Session is closed".

Refs 07720e3080047c765211f6c56d70000b8b3a5043