Ref #48: Allow cluster restart to be resumed
Previously, if the operator would be restarted half-way through the
rolling cluster restarts, the restart would start from the beginning.
Especially on large clusters that could cause a lot of delay, if the
operator would restart right before the last node was restarted.

By raising `kopf.TemporaryError`s and by keeping track of the pods that
need to be restarted, the restart operation can be resumed as often as
necessary.