Preserve keys in annotations, labels, and settings
With Kubernetes version 1.16, keys in objects would only be preserved
if they are either defined in the CustomResourceDefinition or if the
`x-kubernetes-preserve-unknown-fields` key in the object is set to
`true`. The following fields did not specify either keys or the magic
flag, thus never retained any value, rendering them useless.

* `.spec.cluster.settings`
* `.spec.nodes.data.*.annotations`
* `.spec.nodes.data.*.labels`
* `.spec.nodes.data.*.settings`
* `.spec.nodes.master.annotations`
* `.spec.nodes.master.labels`
* `.spec.nodes.master.settings`