Use zone attribute instead of fault domain on Azure
When using an AKS cluster created with `az aks create --zones 1 2 3`,
the various K8s nodes in different zones may share the fault domain.
That means, even if they are spread in different zones, and thus have a
high redundancy, that node attribute can't be used for replication
settings in CrateDB. Hence, using the zone attribute.