Ref #48: Allow cluster scale to be resumed
Previously, if the operator would be restarted half-way through the
scaling operation, the scaling would start from the beginning.
Especially on large clusters that could cause a lot of delay.

By raising `kopf.TemporaryError`s and checking for the state of nodes
and replicas, the scaling operator can be resumed as often as necessary.