Allow returning a status from handlers
Handlers have a facility to store their status. That is stored in the "status" attribute of the resource that is being handled (CrateDB in our case). This was not working before, but is quite useful.

As a result of this change, some handlers had to be adapter to not return large representations (i.e. entire K8s resource objects), because these need to be serialiased and stored against the CrateDB resource (which obviously doesn't work).