Rework cluster health metric gathering
Prior to this change, metrics would "hang around" even if a cluster is deleted. This created a problem as deleted clusters would then trigger the LastSeenTooOld alert and 24x7.

The new method instead has a metric gatherer, which only reports metrics for clusters that exist. If a cluster is deleted, it's metrics will disappear in a few minutes and we won't get alerted.